requirements.txt
prefect
gspread
pandas
psycopg2-binary
oauth2client
